+++ src/com/intellij/idea/plugin/hybris/runConfigurations/HybrisJUnitExtension.kt
@@ -24,18 +24,27 @@ import com.intellij.execution.configurations.ParametersList
 import com.intellij.execution.configurations.RunConfigurationBase
 import com.intellij.execution.configurations.RunnerSettings
 import com.intellij.execution.junit.JUnitConfiguration
-import com.intellij.idea.plugin.hybris.common.HybrisConstants.ENV_HYBRIS_DATA_DIR
-import com.intellij.idea.plugin.hybris.common.HybrisConstants.HYBRIS_DATA_DIRECTORY
+import com.intellij.idea.plugin.hybris.common.HybrisConstants.PROPERTY_BUNDLED_SERVER_TYPE
+import com.intellij.idea.plugin.hybris.common.HybrisConstants.PROPERTY_HYBRIS_BIN_DIR
+import com.intellij.idea.plugin.hybris.common.HybrisConstants.PROPERTY_HYBRIS_BOOTSTRAP_BIN_DIR
+import com.intellij.idea.plugin.hybris.common.HybrisConstants.PROPERTY_HYBRIS_CONFIG_DIR
+import com.intellij.idea.plugin.hybris.common.HybrisConstants.PROPERTY_HYBRIS_DATA_DIR
+import com.intellij.idea.plugin.hybris.common.HybrisConstants.PROPERTY_HYBRIS_LOG_DIR
+import com.intellij.idea.plugin.hybris.common.HybrisConstants.PROPERTY_HYBRIS_ROLES_DIR
+import com.intellij.idea.plugin.hybris.common.HybrisConstants.PROPERTY_HYBRIS_TEMP_DIR
+import com.intellij.idea.plugin.hybris.common.HybrisConstants.PROPERTY_PLATFORMHOME
+import com.intellij.idea.plugin.hybris.common.HybrisConstants.PROPERTY_STANDALONE_JAVAOPTIONS
 import com.intellij.idea.plugin.hybris.common.HybrisConstants.PROPERTY_STANDALONE_JDKMODULESEXPORTS
 import com.intellij.idea.plugin.hybris.project.utils.HybrisRootUtil
 import com.intellij.idea.plugin.hybris.properties.PropertyService
 import com.intellij.idea.plugin.hybris.settings.components.ProjectSettingsComponent
-import com.intellij.openapi.util.io.FileUtil
-import java.io.File
+import com.intellij.openapi.module.Module
+import com.intellij.openapi.module.ModuleManager
+import com.intellij.openapi.project.Project
+import com.intellij.openapi.roots.CompilerModuleExtension
+import com.intellij.openapi.roots.ModuleRootManager
+import com.intellij.openapi.roots.OrderEnumerator
 import java.util.*
-import java.util.regex.Pattern
-
-private val PATTERN = Pattern.compile("\"")
 
 class HybrisJUnitExtension : RunConfigurationExtension() {
 
@@ -50,31 +59,99 @@ class HybrisJUnitExtension : RunConfigurationExtension() {
 
         addVmParameterIfNotExist(vmParameters, "-ea")
 
-        if (vmParameters.parameters.none { it.startsWith("-Dplatformhome=") }) {
-            HybrisRootUtil.findPlatformRootDirectory(project)
-                ?.path
-                ?.let { vmParameters.add("-Dplatformhome=$it") }
-        }
+        addPlatformHome(vmParameters, project)
+        addJavaRunProperties(project, vmParameters)
+        addHybrisEnvProperties(project, vmParameters)
 
-        if (!params.env.containsKey(ENV_HYBRIS_DATA_DIR)) {
-            val settings = ProjectSettingsComponent.getInstance(project).state
-            val hybrisDataDirPath = FileUtil.toCanonicalPath(
-                "${project.basePath}/${settings.hybrisDirectory}/$HYBRIS_DATA_DIRECTORY"
-            )
+        enhanceClassPath(params, project)
+    }
 
-            if (File(hybrisDataDirPath).exists()) {
-                params.addEnv(ENV_HYBRIS_DATA_DIR, hybrisDataDirPath)
+    private fun enhanceClassPath(params: JavaParameters, project: Project) {
+        val classPathEntries = HashSet<String>()
+
+        val modules: Array<Module> = ModuleManager.getInstance(project).modules
+        for (module in modules) {
+            // Get the module's output paths (both production and test)
+            val moduleRootManager = ModuleRootManager.getInstance(module)
+
+            // Get the compiler output paths for production and test
+            val productionOutput = moduleRootManager.getModuleExtension(CompilerModuleExtension::class.java)?.compilerOutputPath
+            val testOutput = moduleRootManager.getModuleExtension(CompilerModuleExtension::class.java)?.compilerOutputPathForTests
+
+            // Add the output paths to the classpath
+            if (productionOutput != null && classPathEntries.add(productionOutput.path)) {
+                params.classPath.add(productionOutput.path)
+            }
+            if (testOutput != null && classPathEntries.add(testOutput.path)) {
+                params.classPath.add(testOutput.path)
+            }
+
+            // **Add module dependencies to classpath**
+            OrderEnumerator.orderEntries(module).recursively().classes().roots.forEach {
+                val path = it.presentableUrl
+                if (classPathEntries.add(path)) {
+                    params.classPath.add(it)
+                }
             }
         }
+    }
+
+    private fun addPlatformHome(vmParameters: ParametersList, project: Project) {
+        val platforhomePrefix = "-D$PROPERTY_PLATFORMHOME="
+        if (vmParameters.parameters.none { it.startsWith(platforhomePrefix) }) {
+            HybrisRootUtil.findPlatformRootDirectory(project)
+                ?.path
+                ?.let { vmParameters.add("$platforhomePrefix=$it") }
+        }
+    }
+
+    private fun addJavaRunProperties(project: Project, vmParameters: ParametersList) {
+        PropertyService.getInstance(project)
+            ?.findProperty(PROPERTY_STANDALONE_JAVAOPTIONS)
+            ?.let { property -> StringTokenizer(property.trim { it <= ' ' }) }
+            ?.let {
+                while (it.hasMoreTokens()) {
+                    val newParam = sanitizeParameter(it.nextToken())
+                    addVmParameterIfNotExist(vmParameters, newParam)
+                }
+            }
 
         PropertyService.getInstance(project)
             ?.findProperty(PROPERTY_STANDALONE_JDKMODULESEXPORTS)
             ?.let { property -> StringTokenizer(property.trim { it <= ' ' }) }
             ?.let {
                 while (it.hasMoreTokens()) {
-                    val newParam = PATTERN.matcher(it.nextToken()).replaceAll("")
+                    val newParam = sanitizeParameter(it.nextToken())
                     addVmParameterIfNotExist(vmParameters, newParam)
                 }
+            }
+
+        PropertyService.getInstance(project)
+            ?.findProperty(PROPERTY_BUNDLED_SERVER_TYPE)
+            ?.let {
+                addVmParameterIfNotExist(vmParameters, "-D$PROPERTY_BUNDLED_SERVER_TYPE=\"$it\"")
+            }
+    }
+
+    private fun sanitizeParameter(param: String): String {
+        return param.replace("\"", "")
+    }
+
+    private fun addHybrisEnvProperties(project: Project, vmParameters: ParametersList) {
+        addPropertyToVmParameter(project, vmParameters, PROPERTY_HYBRIS_BIN_DIR);
+        addPropertyToVmParameter(project, vmParameters, PROPERTY_HYBRIS_TEMP_DIR);
+        addPropertyToVmParameter(project, vmParameters, PROPERTY_HYBRIS_ROLES_DIR);
+        addPropertyToVmParameter(project, vmParameters, PROPERTY_HYBRIS_LOG_DIR);
+        addPropertyToVmParameter(project, vmParameters, PROPERTY_HYBRIS_BOOTSTRAP_BIN_DIR);
+        addPropertyToVmParameter(project, vmParameters, PROPERTY_HYBRIS_DATA_DIR);
+        addPropertyToVmParameter(project, vmParameters, PROPERTY_HYBRIS_CONFIG_DIR);
+    }
+
+    private fun addPropertyToVmParameter(project: Project, vmParameters: ParametersList, propertyKey: String) {
+        PropertyService.getInstance(project)
+            ?.findProperty(propertyKey)
+            ?.let {
+                addVmParameterIfNotExist(vmParameters, "-D$propertyKey=$it")
             }
     }